requirements.txt
langchain_openai
langchain-core
langchain-community
langserve
openai
python-dotenv
streamlit
langchain_community
langchain-ollama

langserve
fastapi
uvicorn
sse_starlette
